Tell cmake where I put FMOD v4.26.36, tell it to not treat implicit integer narrowing as an error
- I got the matching version of FMOD from https://github.com/alexey-lysiuk/fmodex-zdoom
  FMOD no longer makes this available from their own site unless you beg
  them over email.
- Extracted FMOD into ./fmod

Integer narrowing is similar to the situation with uint32 in an earlier
fix, but there were a lot of them and felt that instead of wrapping all
of them in an explicit type cast that I would just disable the error in
this case.

I also couldn't get the vendored version of SDL working, but installing
sdl1-devel let cmake find it without issue.